Code refactor, file clean up and docs improvement
