+++ README.md
@@ -6,65 +6,47 @@ Frontend: React JS, Next JS, Typescript, Tailwind CSS and ShadcnUI
 Backend & Infra: Supabase, Vercel
 Libraries which have small impact may not be listed here or in the project data listing
 
-# Additional Repositories
+## Additional Repositories
 
 There are other repositories might work as support for this one, containing a different set of features:
 
 - [Web Dev Explorer Node JS](https://github.com/Sergie-AGA/web-dev-explorer-nodejs) - Handles additional backend features, such as:
   Node JS, Typescript, Fastify, PostgreSQL, among many others
 
-# Running the project
+## Running the project
 
 This is mainly a Next JS project. Node 18.17+ is recommended for smooth functionality.
 All you should need to run this is installing the modules and running the dev script list on the package.json.
 
-# Documentation
+## Documentation
 
-# Global Overview
+For a more in-depth documentation on the different features and elements being used here, please refer to the following documentation:
 
-## Folder Structure
+- [💻 Application Overview](docs/application-overview.md)
+- [🏠 Homepage](docs/homepage.md)
+- [📦 Product List POC](docs/product-list-poc.md)
+- [🌐 World of CSS](docs/world-of-css.md)
+- [🗂️ Reusable Backend](docs/reusable-backend.md)
 
-## Features
+## Project Overview
 
-## Components
-
-## Config
-
-## Hooks
-
-## Lib
-
-## Services
-
-## Store
-
-## Styles
-
-## Utils
-
-## Tailwind Setup
-
-# Projects
-
-## Project 1 - Product List POC
+### Project 1 - Product List POC
 
 This project experiments a proof of concept of a store, making use of Supabase to handle the backend and modern React ecosystem to render a Product List where you can add items to a cart and trigger discounts based on specific conditions. This was used by me to sell some of my items in the past.
 Frontend: Next JS, React JS, JavaScript, Typescript, Tailwind CSS, ShadcnUI, Zustand, yet-another-react-lightbox, react-photo-album
 Backend: Supabase
 
-## Project 2 - World of CSS Effects
+### Project 2 - World of CSS Effects
 
 A place where diverse CSS tricks and techniques are used to achieve unique visual effects. This is a place where you can find reusable CSS snippets ready to be copied and used in a different project, including examples of usage.
 Frontend: Next JS, React JS, JavaScript, Typescript, Tailwind CSS, ShadcnUI, Zustand
 
-## Project 3 - Reusable Backend
+### Project 3 - Reusable Backend
 
 This project aggregates links to a few different GitHub repositories which contain a reusable modular backend projects in diverse programming languages and can be used as templates for new applications. It currently includes projects in Node JS, PHP and C#.
 Frontend: Next JS, React JS, JavaScript, Typescript, Tailwind CSS, ShadcnUI, Zustand
 Backend: JavaScript, Node JS, Typescript, Fastify, PostgreSQL, Vitest, Supertest, Prisma, Docker, Zod, PHP, Laravel, C#, .NET, ASP.NET,
 
-## In Progress - State Mock Endpoint
+### In Progress - State Mock Endpoint
 
-# Possible improvements
-
-This project could benefit from having Storybook to improve how reusable components can be handled
+In Progress...
