+++ package.json
@@ -55,10 +55,13 @@
     "zustand": "^4.5.0"
   },
   "devDependencies": {
+    "@testing-library/react": "^16.0.1",
     "@types/node": "^20.2.5",
     "@types/react": "^18.2.8",
     "@types/react-dom": "^18.2.4",
     "@types/uuid": "^9.0.8",
+    "cypress": "^13.15.1",
+    "jest": "^29.7.0",
     "postcss-nesting": "^12.0.2",
     "typescript": "^5.1.3"
   }
