+++ src/components/Menu/GlobalMenu.tsx
@@ -11,7 +11,7 @@ import { IconInfoSquare } from "@tabler/icons-react";
 import { useState } from "react";
 import { cn } from "@/utils/utils";
 import AboutModal from "../../features/homepage/components/AboutModal";
-import { modalType, useUIStore } from "@/store/useUIStore";
+import { modalType, useUIStore } from "@/features/homepage/store/useUIStore";
 
 export default function GlobalMenu() {
   const [isOpen, setIsOpen] = useState(false);
