+++ src/features/state-mock-endpoint/components/Menus/SettingsMenu.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import BlurredText from "@/components/Global/Blur/BlurredText";
+import BlurredText from "@/components/Blur/BlurredText";
 import { Button } from "@/components/Shadcn-ui/button";
 import { useUserContext } from "../../context/UserContext";
 import { Separator } from "@/components/Shadcn-ui/separator";
